# Array.vindex performance improvements