--- a/dask/array/tests/test_array_core.py
+++ b/dask/array/tests/test_array_core.py
@@ -3083,7 +3083,7 @@ def test_point_slicing_with_full_slice():
         result = d.vindex[tuple(slc)]
 
         # Rotate the expected result accordingly
-        axis = _get_axis(ind)
+        axis = _get_axis([a for a, i in enumerate(ind) if i is not None])
         expected = _vindex_transpose(x[tuple(slc)], axis)
 
         assert_eq(result, expected)
@@ -3171,18 +3171,6 @@ def test_vindex_errors():
     pytest.raises(IndexError, lambda: d.vindex[[0], [-6]])
     with pytest.raises(IndexError, match="does not support indexing with dask objects"):
         d.vindex[[0], [0], da.array([0])]
-
-
-def test_vindex_merge():
-    from dask.array.core import _vindex_merge
-
-    locations = [1], [2, 0]
-    values = [np.array([[1, 2, 3]]), np.array([[10, 20, 30], [40, 50, 60]])]
-
-    assert (
-        _vindex_merge(locations, values)
-        == np.array([[40, 50, 60], [1, 2, 3], [10, 20, 30]])
-    ).all()
 
 
 def test_vindex_identity():