This is a patch to `Array.vindex` that aims to:
1. Cut down on how long it takes to build the task graph
2. Reduce memory usage in some cases by avoiding copies of the index arrays

For execute time, I used the following code as a simple benchmark:
```python
import dask.array as da
import numpy as np
arr_1 = da.ones([6000, 6000], chunks=[3000, 3000])  # 36M elements in 4 chunks
idx = np.repeat(np.arange(0, 6000, 6), 1000)  # 1M index points
arr_2 = arr_1.vindex[idx, idx[::-1]]
```

Before this change the last line took 2.47s of wall clock time to execute. With this change it takes 22.1ms.

Regarding memory usage, I was motivated by my use case which involves accessing NetCDF files via Xarray and using `xarray.Dataset.isel`, which under the hood calls `da.Array.vindex` on many arrays using the same index values. The current `vindex` implementation always stores copies of the index arrays in the task graph, so an `isel` call could cause many index array copies to accumulate. This change stores the passed-in index arrays directly in the task graph, if possible.

This implementation essentially works the same way as the existing one, it just handles the bookkeeping data structures a bit differently. One consequence is that some of the index processing that used to happen when building the graph now happens at graph execution time. Nonetheless, in my testing graph execution performance also seems to have improved a bit: wall time for `arr_2.compute()` from the snippet above dropped from 632ms to 118ms.

I don't have much experience with Dask internals, but I hope this is a workable solution. Also, a couple questions:
1. Is it ok to store the passed-in index arrays directly in the task graph like this? If the caller goes on to mutate these arrays, this could cause surprising results. If it's not ok it'd be easy to change the code to make a copy and still benefit from the speedup. But I suspect it is ok, since a similar situation results just from calling `da.from_array` on a NumPy array.
2. From profing `vindex` I can see that more than half its time is now being spent hashing the index arrays in order to assign key names to graph nodes. Are there any alternatives for generating keys in this type of situation?